Project New/ai-insurance-agent-mvp/requirements.txt
streamlit>=1.35.0
psycopg2-binary>=2.9.6
sqlalchemy>=2.0.30
python-dotenv>=1.0.1
requests>=2.32.0
anthropic>=0.28.0
pandas>=2.2.2
plotly>=5.22.0
apscheduler>=3.10.4
pydantic>=2.9.0
httpx>=0.27.0
pytz>=2024.1
phonenumbers>=8.13.0
babel>=2.14.0
streamlit-autorefresh>=1.0.1
fastapi>=0.111.0
uvicorn>=0.30.0